+++ src/main/java/com/rootnext/jurassicpark/model/SpeciesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by sanjoy on 6/29/15.
  */
-public interface SpeciesRepository extends CrudRepository<Species, Integer>{
+public interface SpeciesRepository extends CrudRepository<Species, Long>{
 }
